requirements.txt
Flask==2.1.0
tabula-py==2.2.0
pandas==1.3.3
openpyxl==3.0.9
